requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
scipy>=1.7.0
cvxpy>=1.1.0
osqp>=0.6.0
pyyaml>=5.4.0